Make all actions asynchronous by using Future and the global EC
All JDBC actions were executed sequential because they all waited for the JDBC action to finish. This blocked all following executions. The invocations to JDBC are now wrapped inside a Future{} so that the tests can continue normally.
Introduce a BlockLatchAction for unit tests to be able to wait for the futures.

Closes #2